# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rhodes Analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorMessage {
  error: string;
}

export interface Site {
  id: number;
  name: string;
  domain: string;
  trackingId: string;
  createdAt: string;
}

export interface SiteOverview {
  id: number;
  name: string;
  domain: string;
  trackingId: string;
  createdAt: string;
  visitors24h: number;
  pageviews24h: number;
  liveVisitors: number;
  /** Hourly pageview counts for the last 24 hours (oldest first) */
  sparkline: number[];
}

export interface SiteInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  domain: string;
}

export interface SiteUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  domain?: string;
}

export interface EventInput {
  trackingId: string;
  path: string;
  /** @nullable */
  referrer?: string | null;
  visitorId: string;
  /** @nullable */
  userAgent?: string | null;
  /**
     * ISO 3166-1 alpha-2 country code
     * @nullable
     */
  country?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
}

export interface CollectResult {
  accepted: boolean;
}

export interface RecentEvent {
  id: number;
  path: string;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  countryName?: string | null;
  /** @nullable */
  browser?: string | null;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  deviceType?: string | null;
  occurredAt: string;
}

export interface StatsSummary {
  visitors: number;
  pageviews: number;
  /** Percentage 0-100 */
  bounceRate: number;
  avgDurationSeconds: number;
  /**
     * Percent change vs previous period; null when no prior data
     * @nullable
     */
  visitorsChange: number | null;
  /** @nullable */
  pageviewsChange: number | null;
  /** @nullable */
  bounceRateChange: number | null;
  /** @nullable */
  avgDurationChange: number | null;
  /** Unique visitors in the last 5 minutes */
  liveVisitors: number;
}

export interface TimeseriesPoint {
  /** ISO timestamp for the start of the bucket */
  bucket: string;
  visitors: number;
  pageviews: number;
}

export interface PageStat {
  path: string;
  visitors: number;
  pageviews: number;
}

export interface ReferrerStat {
  /** Referrer source, "Direct" when none */
  referrer: string;
  visitors: number;
}

export interface BreakdownItem {
  name: string;
  visitors: number;
  percentage: number;
}

export interface DeviceBreakdown {
  browsers: BreakdownItem[];
  operatingSystems: BreakdownItem[];
  deviceTypes: BreakdownItem[];
}

export interface CountryStat {
  /** ISO 3166-1 alpha-2 code */
  country: string;
  countryName: string;
  visitors: number;
}

export interface TopSite {
  id: number;
  name: string;
  domain: string;
  visitors24h: number;
}

export interface Overview {
  totalSites: number;
  visitors24h: number;
  pageviews24h: number;
  liveVisitors: number;
  topSite: TopSite | null;
}

export type RangeParameter = typeof RangeParameter[keyof typeof RangeParameter];


export const RangeParameter = {
  '24h': '24h',
  '7d': '7d',
  '30d': '30d',
  '90d': '90d',
} as const;

export type GetRecentEventsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type GetStatsSummaryParams = {
range?: RangeParameter;
};

export type GetStatsTimeseriesParams = {
range?: RangeParameter;
};

export type GetStatsPagesParams = {
range?: RangeParameter;
};

export type GetStatsReferrersParams = {
range?: RangeParameter;
};

export type GetStatsDevicesParams = {
range?: RangeParameter;
};

export type GetStatsCountriesParams = {
range?: RangeParameter;
};