# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rhodes Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List sites with 24h stats and sparkline
      responses:
        "200":
          description: Sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteOverview"
    post:
      operationId: createSite
      tags: [sites]
      summary: Register a new site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get a site
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "200":
          description: Site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    patch:
      operationId: updateSite
      tags: [sites]
      summary: Update a site
      parameters:
        - $ref: "#/components/parameters/SiteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete a site and its events
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /collect:
    post:
      operationId: collectEvent
      tags: [events]
      summary: Ingest a pageview event from the tracking snippet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /sites/{id}/events/recent:
    get:
      operationId: getRecentEvents
      tags: [events]
      summary: Recent events feed for a site
      parameters:
        - $ref: "#/components/parameters/SiteId"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Recent events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentEvent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
  parameters:
    SiteId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Site:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        domain:
          type: string
        trackingId:
          type: string
        createdAt:
          type: string
      required: [id, name, domain, trackingId, createdAt]
    SiteOverview:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        domain:
          type: string
        trackingId:
          type: string
        createdAt:
          type: string
        visitors24h:
          type: number
        pageviews24h:
          type: number
        liveVisitors:
          type: number
        sparkline:
          type: array
          items:
            type: number
          description: Hourly pageview counts for the last 24 hours (oldest first)
      required:
        [
          id,
          name,
          domain,
          trackingId,
          createdAt,
          visitors24h,
          pageviews24h,
          liveVisitors,
          sparkline,
        ]
    SiteInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        domain:
          type: string
          minLength: 1
      required: [name, domain]
    SiteUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        domain:
          type: string
          minLength: 1
    EventInput:
      type: object
      properties:
        trackingId:
          type: string
        path:
          type: string
        referrer:
          type: ["string", "null"]
        visitorId:
          type: string
        userAgent:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
          description: ISO 3166-1 alpha-2 country code
        durationSeconds:
          type: ["number", "null"]
      required: [trackingId, path, visitorId]
    CollectResult:
      type: object
      properties:
        accepted:
          type: boolean
      required: [accepted]
    RecentEvent:
      type: object
      properties:
        id:
          type: number
        path:
          type: string
        referrer:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        countryName:
          type: ["string", "null"]
        browser:
          type: ["string", "null"]
        os:
          type: ["string", "null"]
        deviceType:
          type: ["string", "null"]
        occurredAt:
          type: string
      required: [id, path, occurredAt]
...