# lib/api-zod/src/generated/types/recentEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rhodes Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface RecentEvent {
  id: number;
  path: string;
  /** @nullable */
  referrer?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  countryName?: string | null;
  /** @nullable */
  browser?: string | null;
  /** @nullable */
  os?: string | null;
  /** @nullable */
  deviceType?: string | null;
  occurredAt: string;
}