# lib/api-zod/src/generated/types/siteOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rhodes Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteOverview {
  id: number;
  name: string;
  domain: string;
  trackingId: string;
  createdAt: string;
  visitors24h: number;
  pageviews24h: number;
  liveVisitors: number;
  /** Hourly pageview counts for the last 24 hours (oldest first) */
  sparkline: number[];
}