# lib/api-zod/src/generated/types/siteUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Rhodes Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  domain?: string;
}